src/buvis/pyproject.toml
[tool.poetry]
name = "buvis"
version = "0.1.0"
description = ""
authors = ["Tomáš Bouška <tomas@buvis.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
rich = "^13.8.0"
tzlocal = "^5.2"
inflection = "^0.5.1"
doogat-core = "^0.1.0"


[tool.poetry.group.dev.dependencies]
doogat-core = {path = "../../../../doogat/doogat-core", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
